refactor code locations
