+++ src/main/java/pkg/deepCurse/pandora/core/mixins/client/GameRendererMixin.java
@@ -17,8 +17,8 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.GameRenderer;
 import net.minecraft.client.render.LightmapTextureManager;
 import net.minecraft.client.util.math.MatrixStack;
-import pkg.deepCurse.pandora.core.interfaces.LightmapAccess;
-import pkg.deepCurse.pandora.tools.DarknessTools;
+import pkg.deepCurse.pandora.core.util.interfaces.LightmapAccess;
+import pkg.deepCurse.pandora.core.util.tools.DarknessTools;
 
 @Mixin(GameRenderer.class)
 public class GameRendererMixin {
