+++ src/main/java/pkg/deepCurse/pandora/core/util/tools/PandoraTools.java
@@ -1,4 +1,4 @@
-package pkg.deepCurse.pandora.tools;
+package pkg.deepCurse.pandora.core.util.tools;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
