+++ src/main/resources/fabric.mod.json
@@ -22,7 +22,7 @@
 			"pkg.deepCurse.pandora.core.Pandora"
 		],
 		"modmenu": [
-			"pkg.deepCurse.pandora.core.integrations.ModMenuIntegration"
+			"pkg.deepCurse.pandora.core.util.integrations.ModMenuIntegration"
 		]
 	},
 	"mixins": [
